generated_advanced/requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.12.1
psycopg2-binary==2.9.10
python-dotenv==1.0.0
